requirements.txt
asgiref==3.4.1
Django==3.2.5
django-rest-auth==0.9.5
django-summernote==0.8.11.6
djangorestframework==3.12.4
img==2.5
mysqlclient==2.0.3
Pillow==8.3.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
